requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
pytz>=2025.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
